refactor(project structure): move trait definition to separate package
