+++ operators/poc-secondary-resource-management/Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
-members = ["demo"]
+members = ["demo", "lib"]
 resolver = "2"
