+++ operators/poc-secondary-resource-management/demo/Cargo.toml
@@ -14,6 +14,7 @@ name = "crdgen"
 path = "src/crdgen.rs"
 
 [dependencies]
+lib = { path = "../lib" }
 kube = { version = "0.97.0", features = ["runtime", "derive", "client"] }
 k8s-openapi = { version = "0.23.0", features = ["v1_31", "schemars"] }
 tokio = { version = "1.41.1", features = ["full"] }
@@ -21,7 +22,6 @@ serde = { version = "1.0.215", features = ["derive"] }
 kube-derive = "0.97.0"
 schemars = "0.8.21"
 serde_json = "1.0.133"
-thiserror = "2.0.3"
 futures = "0.3.31"
 serde_yaml = "0.9.34"
 log = "0.4.22"
