+++ operators/poc-secondary-resource-management/demo/src/controller.rs
@@ -1,8 +1,7 @@
 use std::sync::Arc;
 use std::time::Duration;
 
-use crate::crd::{self};
-use crate::error::{Error, Result};
+use crate::crd;
 use crate::reconcile_sec_res;
 use futures::StreamExt;
 use k8s_openapi::api::apps;
@@ -11,6 +10,7 @@ use kube::core::object::HasSpec;
 use kube::runtime::{watcher, Controller};
 use kube::{runtime::controller::Action, Api};
 use kube::{Client, ResourceExt};
+use lib::error::{Error, Result};
 
 const LABEL_KUBERNETES_REPLSET: &str = "app.kubernetes.io/replset";
 const LABEL_KUBERNETES_COMPONENT: &str = "app.kubernetes.io/component";
