+++ operators/poc-secondary-resource-management/demo/src/crdgen.rs
@@ -1,5 +1,4 @@
 pub mod crd;
-pub mod error;
 
 use kube::CustomResourceExt;
 use std::fs::File;
