+++ operators/poc-secondary-resource-management/lib/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "lib"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+kube = { version = "0.97.0", features = ["runtime", "derive", "client"] }
+k8s-openapi = { version = "0.23.0", features = ["v1_31", "schemars"] }
+serde = { version = "1.0.215", features = ["derive"] }
+thiserror = "2.0.3"
+futures = "0.3.31"
+log = "0.4.22"
